Feature/23515 aplicar save localstorage test

+++ Src/GestaoAvaliacao.Entities/Alternative.cs
@@ -28,12 +28,13 @@ namespace GestaoAvaliacao.Entities
         public int ItemOrder { get; set; }
 
         [NotMapped]
-        public int Item_Id { get; set; }
+        public long Item_Id { get; set; }
 
         [NotMapped]
         public bool Selected { get; set; }
 
         [NotMapped]
         public string NumerationSem { get; set; }
+
     }
 }
